requirements.txt
asttokens==2.2.1
autopep8==2.0.2
backcall==0.2.0
cachetools==5.3.0
certifi==2022.12.7
charset-normalizer==3.1.0
comm==0.1.3
debugpy==1.6.7
decorator==5.1.1
executing==1.2.0
gevent==22.10.2
google-api-core==2.11.0
google-api-python-client==2.86.0
google-auth==2.17.3
google-auth-httplib2==0.1.0
googleapis-common-protos==1.59.0
greenlet==2.0.2
grpcio==1.54.0
grpcio-status==1.54.0
httplib2==0.22.0
idna==3.4
ipykernel==6.22.0
ipython==8.12.1
jedi==0.18.2
jupyter_client==8.2.0
jupyter_core==5.3.0
kafka-python==2.0.2
matplotlib-inline==0.1.6
nest-asyncio==1.5.6
numpy==1.24.3
packaging==23.1
pandas==2.0.1
parso==0.8.3
pexpect==4.8.0
pickleshare==0.7.5
platformdirs==3.5.0
prompt-toolkit==3.0.38
protobuf==4.22.3
psutil==5.9.5
psycopg2-binary==2.9.6
ptyprocess==0.7.0
pure-eval==0.2.2
py4j==0.10.9.5
pyarrow==11.0.0
pyasn1==0.5.0
pyasn1-modules==0.3.0
pycodestyle==2.10.0
Pygments==2.15.1
pyparsing==3.0.9
pyspark==3.3.2
python-dateutil==2.8.2
pytz==2023.3
pyzmq==25.0.2
requests==2.29.0
rsa==4.9
six==1.16.0
stack-data==0.6.2
tornado==6.3.1
traitlets==5.9.0
tzdata==2023.3
uritemplate==4.1.1
urllib3==1.26.15
wcwidth==0.2.6
websocket==0.2.1
websocket-client==1.5.1
zope.event==4.6
zope.interface==6.0
